Reorder for new structure
